maintenant on a une première version de la souris mais l'image du block ne suit pas la souris en temps réel, elle le suit que quand on click.

+++ src/main/java/modele/ItemBlock.java
@@ -14,8 +14,14 @@ public class ItemBlock extends Item{
     @Override
     public void use() {
        Block bPlace = new Block(this, (MouseHandler.mouseX.getValue()/32) * 32, (MouseHandler.mouseY.getValue()/ 32) * 32, Controleur.terrain);
-       if(Controleur.terrain.placeBlock(bPlace.getHitX(), bPlace.getHitY(), bPlace)){
-           Controleur.player.drop();
+       if (Controleur.terrain.placeBlock(bPlace.getHitX(), bPlace.getHitY(), bPlace)) {
+           if(Controleur.player.getInventory().getCurrItem() != null){
+               Controleur.player.drop();
+           }
+           else if (Controleur.playerMouse.item != null){
+               Controleur.playerMouse.decremente(1);
+           }
+
        }
 
     }
